Can now get through vaccination phase

+++ test/d3/demo/NodeSizeComparator.java
@@ -15,8 +15,8 @@ public class NodeSizeComparator implements Comparator<WebElement>
     @Override
     public int compare(WebElement o1, WebElement o2)
     {
-        int r1 = Integer.parseInt(o1.getAttribute("r"));
-        int r2 = Integer.parseInt(o2.getAttribute("r"));
+        double r1 = Double.parseDouble(o1.getAttribute("r"));
+        double r2 = Double.parseDouble(o2.getAttribute("r"));
         if (r1 < r2)
             return -1;
         else if (r1 == r2)
